global: runner module, separated tests for rdm/videos
