+++ cds_migrator_kit/base_minter.py
@@ -6,17 +6,17 @@
 # the terms of the MIT License; see LICENSE file for more details.
 """Minter configuration for CDS Migrator Kit."""
 
-import importlib
 import warnings
+
+from cds_migrator_kit.import_utils import import_module
 
 # Default: No minter
 selected_minter = None
 
 # Check if `rdm` is installed and set the minter
-try:
-    importlib.import_module("cds_rdm.__init__")
+if import_module("cds_rdm.__init__"):
     from cds_rdm.minters import legacy_recid_minter as selected_minter
-except ImportError:
+else:
     warnings.warn(
         "No valid PID minter found. Ensure `rdm` is installed.", RuntimeWarning
     )
