+++ cds_migrator_kit/runner/__init__.py
@@ -5,4 +5,4 @@
 # cds-migrator-kit is free software; you can redistribute it and/or modify it under
 # the terms of the MIT License; see LICENSE file for more details.
 
-"""Extract module."""
+"""Runner module."""
