+++ cds_migrator_kit/runner/runner.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright (C) 2022 CERN.
+# Copyright (C) 2022-2025 CERN.
 #
-# Invenio-RDM-Migrator is free software; you can redistribute it and/or modify
+# cds-migrator-kit is free software; you can redistribute it and/or modify
 # it under the terms of the MIT License; see LICENSE file for more details.
 
 """InvenioRDM migration streams runner."""
