+++ cds_migrator_kit/videos/weblecture_migration/cli.py
@@ -13,7 +13,7 @@ import click
 from flask import current_app
 from flask.cli import with_appcontext
 
-from cds_migrator_kit.rdm.runner import Runner
+from cds_migrator_kit.runner.runner import Runner
 from cds_migrator_kit.videos.weblecture_migration.streams import RecordStreamDefinition
 
 cli_logger = logging.getLogger("migrator")
