+++ tests/cds-rdm/data/summer_note.json
