+++ tests/cds-rdm/test_json_translation_rules.py
